requirements.txt
langchain-ollama
langchain-core
llama-index-core
